only use `List` for arguments, never `Collection` [api]
As in OSC, the messages use the `List` logic
(there can be duplicates and order matters),
it may only confuse the user of the library,
if he sees that he can supply for example a `Set`.
The user might get insecure of whether the duplicate entries are allowed,
and whether the order of the arguments is maintained after a serialize
and re-parse cycle.