discard any message with an unknown argument type tag
This is suggested behavior; to be found in the OSC specification 1.0.
We currently print an error message on `stderr` for now,
to give the user at least a slight chance to find out what went wrong.
This would be a good reason to use a logging library,
even if just the Java internal one.