switch to Java 8
As of January 2019, it is not even possible to download Java 7
from Oracle (without special priviledges) anymore,
and it is also not available anymore in recent Linux package repositories.

If you still need recent JavaOSC changes,
yet compatible with Java 6 or 7,
please compile it yourself or ask for help by opening an issue.

NOTE Java 8 was released in March 2014.