remove tests about default ports
I decided to remove some of the getter methods from OSCPort/OSCPortIn/OSCPortOut
that concern themselves with transport-level implementation details. This
resulted in compiler errors in the testPorts test case, as getRemoteAddress no
longer exists.

I thought about this, and I think that removing getRemoteAddress makes sense in
the context of real world usage. I don't think it makes sense to keep it just to
facilitate having these tests, which are only testing that the default values we
specified are the values you get when you don't provide a value.

I'm open to feedback on this - happy to add this back if needed!