simplify serializer to return a byte array instead of writing to a passed-in ByteBuffer
Prior to this commit, the serializer methods operated on a pre-allocated
ByteBuffer, then after all the mutation took place, you could either use the
ByteBuffer as-is to write to the DatagramChannel for UDP, or pull out the
underlying byte array of the ByteBuffer and use that to write to the output
stream for TCP.

Problem #1: Side effects / mutation makes code difficult to understand and
reason about because your brain has to act like an interpreter, stepping through
all the instructions and attempting to keep track of the state of all the
mutable things.

Problem #2: ByteBuffers and byte arrays are fixed-length, but there is no limit
to the size of the content of a TCP stream. So, passing in a pre-allocated
ByteBuffer or byte array won't work. Given an OSCPacket of arbitrarily large
size, we want to convert it to a byte array and _then_ worry about writing the
bytes to an OutputStream (TCP) or DatagramChannel (UDP).

This commit solves both of these problems by turning the methods in the
implementation of OSCSerializer into pure functions with the signature:

    Object -> byte[]

The contract of OSCSerializer now is that it "serializes" (i.e. returns a byte
array), whereas before, the contract was that it "writes" to a ByteBuffer.