add development-only log4j config file
While developing, I noticed that the OSCPortIn listener thread was logging
errors, but they weren't visible due to the lack of a log4j.properties file,
i.e. log4j isn't configured.

Given that JavaOSC is a library, it probably makes sense for the project NOT to
include a log4j.properties file, because the log4j configuration might leak out
into projects that consume this library. (This is my assumption; maybe I'm
wrong.)

log4j configuration should be controlled by consumers of this library, but when
we're developing JavaOSC and running tests, etc., it is useful to be able to
configure logging. As such, I've created a log4j-dev.properties file that you
can opt into using by including the -Dlog4j.configuration=log4j-dev.properties
property when you invoke mvn.

Worth noting: we have checkstyle configured to enforce a header based on
java_header_regex_template.txt, but the comment syntax in that template is
Java-specific. So, this commit also adjusts our checkstyle configuration so that
we're only enforcing that rule for *.java files.