OSCPort{In,Out}Builder: use a wildcard port for {remote,local} when unspecified
The most important thing this fixes is that when you create an OSCPortOut, you
can use setRemotePort and not specify a local port. If you do specify a local
port, it binds on that port, which is undesirable for TCP because it occupies
that port so that nothing else can listen on it.