bugfix: don't terminate/align (i.e. add extra padding to) the packet as a whole
I was misunderstanding when we should terminate and align. It's basically just
for things like strings (including type tags) and blobs, which can end up having
a byte size that isn't divisible by 4. If we handle those properly, then an
OSCBundle or an OSCMessage should always have a byte size divisible by 4, and we
don't need to add additional padding at the end. Having the additional padding
at the end trips up both the JavaOSC server the go-osc server.